mainapp/serializers.py
from rest_framework import serializers
from .models import Test, Question, SpecialistCategory, Specialist, UserTestResult


class TestSerializer(serializers.ModelSerializer):
    class Meta:
        model = Test
        fields = '__all__'

class QuestionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Question
        fields = '__all__'

class SpecialistCategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = SpecialistCategory
        fields = '__all__'

class SpecialistSerializer(serializers.ModelSerializer):
    class Meta:
        model = Specialist
        fields = '__all__'

class UserTestResultSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserTestResult
        fields = '__all__'