requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.5.0
async-timeout==4.0.3
attrs==21.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
Django==3.2.7
django-environ==0.11.2
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
environs==9.5.0
exceptiongroup==1.1.3
eyeD3==0.8.10
filelock==3.6.0
frozenlist==1.4.0
grpcio==1.30.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
loguru==0.7.2
louis==3.20.0
macaroonbakery==1.3.1
magic-filter==1.0.11
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.20.1
more-itertools==8.10.0
multidict==6.0.4
nemo-emblems==5.6.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
onboard==1.4.1
packaging==21.3
PAM==0.4.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.12.4
psutil==5.9.0
psycopg2==2.9.7
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pydantic==2.3.0
pydantic_core==2.6.3
pyelftools==0.27
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyTelegramBotAPI==4.12.0
python-apt==2.4.0+ubuntu2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==0.19.2
python-gnupg==0.4.8
python-magic==0.4.24
python-telegram-bot==20.5
python-xlib==0.29
pytils==0.4.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
SecretStorage==3.3.1
selectolax==0.3.16
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
sqlparse==0.4.2
systemd-python==234
texttable==1.6.4
tinycss2==1.1.1
tldextract==3.1.2
typing_extensions==4.7.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.2.3
xdg==5
xkit==0.0.0
xlrd==1.2.0
yarl==1.9.2
youtube-dl==2021.12.17
zipp==1.0.0
